# openapi-spec.yaml
openapi: 3.0.0
info:
  title: NightGuide API
  version: 0.1.0
servers:
  - url: https://api.nightguide.app
  - url: https://api.dev.nightguide.app
paths:
...
  /venues:
    get:
      summary: Get venues.
      operationId: getVenues
      tags:
        - venues
      parameters:
        - in: query
          name: pageSlug
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: cat
          schema:
            type: array
            items:
              type: string
        - in: query
          name: tag
          schema:
            type: array
            items:
              type: string
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
        - in: query
          name: hasFb
          schema:
            type: boolean
        - in: query
          name: musicType
          schema:
            type: array
            items:
              type: string
        - in: query
          name: visitorType
          schema:
            type: array
            items:
              type: string
        - in: query
          name: paymentMethod
          schema:
            type: array
            items:
              type: string
        - in: query
          name: doorPolicy
          schema:
            type: array
            items:
              type: string
        - in: query
          name: dresscode
          schema:
            type: array
            items:
              type: string
        - in: query
          name: capRange
          schema:
            type: array
            items:
              type: integer
              format: int32
              minimum: 1
              maximum: 7
        - in: query
          name: priceClass
          schema:
            type: array
            items:
              type: integer
              format: int32
              minimum: 1
        - in: query
          name: noEntranceFee
          schema:
            type: boolean
        - in: query
          name: noCoatCheckFee
          schema:
            type: boolean
        - in: query
          name: noBouncers
          schema:
            type: boolean
        - in: query
          name: openTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: terraceTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: kitchenTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: busyTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: dancingTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: bitesTime
          schema:
            type: string
            format: 'date-time'
        - in: query
          name: vipArea
          schema:
            type: boolean
        - in: query
          name: smokingArea
          schema:
            type: boolean
        - in: query
          name: terrace
          schema:
            type: boolean
        - in: query
          name: terraceHeaters
          schema:
            type: boolean
        - in: query
          name: bouncers
          schema:
            type: boolean
        - in: query
          name: kitchen
          schema:
            type: boolean
        - in: query
          name: coatCheck
          schema:
            type: boolean
        - in: query
          name: parking
          schema:
            type: boolean
        - in: query
          name: cigarettes
          schema:
            type: boolean
        - in: query
          name: accessible
          schema:
            type: boolean
        - in: query
          name: query
          schema:
            type: string
            minLength: 2
            maxLength: 255
          description: Search query
        - in: query
          name: sortBy
          schema:
            type: array
            items:
              type: string
          description: Sort with optional order
        - in: query
          name: ids
          schema:
            type: array
            items:
              type: string
        - in: query
          name: exclude
          schema:
            type: array
            items:
              type: string
        - in: query
          name: longitude
          schema:
            type: number
          description: Longitude for distance sort
        - in: query
          name: latitude
          schema:
            type: number
          description: Latitude for distance sort
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
            minimum: 0
          description: Number of items to skip
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50
          description: Number of items to return. Defaults to 20.
        - in: query
          name: fields
          schema:
            type: array
            items:
              type: string
          description: Fields to return. Returns limited set of fields by default.
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Result venues.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Venue'
                  offset:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  totalCount:
                    type: integer
                    format: int32
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a venue.
      operationId: createVenue
      tags:
        - venues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venue'
      responses:
        '201':
          description: Venue created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /venues/{venueId}:
    get:
      summary: Get a venue.
      operationId: getVenue
      tags:
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Venue data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update a venue.
      operationId: updateVenue
      tags:
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venue'
      responses:
        '200':
          description: Venue updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a venue.
      operationId: deleteVenue
      tags:
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /venues/{venueId}/images:
    post:
      summary: Create a venue image.
      operationId: createVenueImage
      tags:
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
          application/json:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
      responses:
        '200':
          description: Image uploaded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /venues/{venueId}/images/{imageId}:
    delete:
      summary: Delete a venue image.
      operationId: deleteVenueImage
      tags:
        - events
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
        - in: path
          name: imageId
          schema:
            type: string
          required: true
          description: Image ID.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /venues/{venueId}/facebook-events:
    put:
      summary: Update a venue's Facebook events.
      operationId: updateFacebookVenueEvents
      tags:
        - events
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Event'
                required:
                  - facebook
                  - dates
      responses:
        '200':
          description: Events updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /venues/{venueId}/generate-ticket-codes:
    post:
      operationId: generateVenueTicketCodes
      tags:
        - venues
      parameters:
        - in: path
          name: venueId
          schema:
            type: string
          required: true
          description: Venue ID.
      responses:
        '200':
          description: Ticket codes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  codes:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /events/{eventId}:
    get:
      summary: Get an event.
      operationId: getEvent
      tags:
        - events
      parameters:
        - in: path
          name: eventId
          schema:
            type: string
          required: true
          description: Event ID.
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Event data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update an event.
      operationId: updateEvent
      tags:
        - events
      parameters:
        - in: path
          name: eventId
          schema:
            type: string
          required: true
          description: Event ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Event updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete an event.
      operationId: deleteEvent
      tags:
        - events
      parameters:
        - in: path
          name: eventId
          schema:
            type: string
          required: true
          description: Event ID.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: The request did not satisfy the requirements
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
            required:
              - success
  schemas:
...
    Image:
      required:
        - id
        - filename
        - filetype
        - filesize
        - url
        - width
        - height
      properties:
        id:
          type: string
          description: UUIDv4 image ID (also part of filename).
        filename:
          type: string
        filetype:
          type: string
          description: Image's MIME type.
        filesize:
          type: number
        url:
          type: string
          description: Serveable image URL.
        width:
          type: number
          description: Image width.
        height:
          type: number
          description: Image height.
        extraData:
          type: object
    Venue:
      type: object
      properties:
        id:
          type: string
        sourceId:
          type: number
          format: int32
        name:
          type: string
        queryText:
          type: string
        pageSlug:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/VenueCategories'
        description:
          $ref: '#/components/schemas/TranslatedObject'
        location:
          type: object
          properties:
            address1:
              type: string
            address2:
              type: string
            postalCode:
              type: string
            city:
              type: string
            country:
              type: string
              minLength: 2
              maxLength: 2
            coordinates:
              type: object
              properties:
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
            googlePlaceId:
              type: string
        website:
          type: string
        facebook:
          type: object
          properties:
            id:
              type: string
            pagesId:
              type: string
        instagram:
          type: object
          properties:
            id:
              type: string
            explorePage:
              type: string
        twitterHandle:
          type: string
        musicTypes:
          type: array
          items:
            $ref: '#/components/schemas/VenueMusicTypes'
        visitorTypes:
          type: array
          items:
            $ref: '#/components/schemas/VenueVisitorTypes'
        dresscode:
          $ref: '#/components/schemas/VenueDresscodes'
        paymentMethods:
          type: array
          items:
            $ref: '#/components/schemas/VenuePaymentMethods'
        doorPolicy:
          type: object
          properties:
            policy:
              $ref: '#/components/schemas/VenueDoorPolicies'
            description:
              $ref: '#/components/schemas/TranslatedObject'
        prices:
          type: object
          properties:
            coke:
              type: number
              format: double
            beer:
              type: number
              format: double
        priceClass:
          type: integer
          format: int32
          minimum: 1
        fees:
          type: object
          properties:
            coatCheck:
              type: number
              format: double
            entrance:
              type: number
              format: double
        entranceFeeRange:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
        capacity:
          type: integer
          format: int32
          minimum: 0
        capacityRange:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
        facilities:
          type: array
          items:
            $ref: '#/components/schemas/VenueFacilities'
        images:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/Image'
        admin:
          type: object
          properties:
            hide:
              type: boolean
    VenueFacilities:
      type: string
      enum:
        [
          'vip_area',
          'smoking_area',
          'bouncers',
          'kitchen',
          'coat_check',
          'parking',
          'cigarettes',
          'accessible',
          'terrace',
          'terrace_heaters',
        ]
    VenueDoorPolicies:
      type: string
      enum: ['moderate', 'strict', 'guestlist']
    VenuePaymentMethods:
      type: string
      enum: ['debit_card', 'credit_card', 'cash']
    VenueDresscodes:
      type: string
      enum: ['chique', 'alternative']
    VenueVisitorTypes:
      type: string
      enum:
        [
          'teenagers',
          'students',
          'young_professionals',
          'middle_aged',
          'seniors',
          'international',
          'locals',
          'lgbtq',
          'varying',
        ]
    VenueMusicTypes:
      type: string
      enum:
        [
          'techno',
          'dance',
          'live',
          'varying',
          'jazz',
          'apres_ski',
          'country',
          'miscellaneous',
          'folk',
          'house',
          'lounge',
          'no_music',
          'hip_hop',
          'rock',
          'classical',
          '80s_90s',
          'disco',
          'latin',
          'electro',
        ]
    VenueCategories:
      type: string
      enum:
        [
          'club',
          'dancing_bar',
          'beach_club',
          'bar',
          'lounge_bar',
          'craft_beer_bar',
          'cocktail_bar',
          'concert_venue',
          'strip_club',
          'blues_bar',
          'casino',
          'comedy_bar',
          'country_bar',
          'hookah_bar',
          'irish_pub',
          'jazz_bar',
          'karaoke_bar',
          'lgbtq_bar',
          'piano_bar',
          'sports_bar',
          'wine_bar',
          'whisky_bar',
        ]
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        queryText:
          type: string
        date:
          $ref: '#/components/schemas/EventDate'
        dates:
          type: array
          items:
            $ref: '#/components/schemas/EventDate'
        description:
          $ref: '#/components/schemas/TranslatedObject'
        location:
          type: object
          required:
            - type
          properties:
            type:
              type: string
              enum: ['address', 'venue']
            address1:
              type: string
            address2:
              type: string
            postalCode:
              type: string
            city:
              type: string
            country:
              type: string
              minLength: 2
              maxLength: 2
            coordinates:
              type: object
              properties:
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
        images:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/Image'
        facebook:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            description:
              type: string
            datesChanged:
              type: boolean
        tickets:
          type: object
          properties:
            provider:
              type: string
            checkoutUrl:
              type: string
            providerData:
              type: object
            products:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
                    format: float
        admin:
          type: object
          properties:
            hide:
              type: boolean
    EventDate:
      type: object
      properties:
        id:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        interestedCount:
          type: number
          format: int32
          nullable: true
        ticketsUrl:
          type: string
        providerEventId:
          type: string
        artists:
          type: array
          items:
            oneOf:
              - type: object
              - type: string
        isHot:
          type: boolean
...
    TranslatedObject:
      type: object
      properties:
        en:
          type: string
        nl:
          type: string
    Error:
      type: object
      properties:
        status:
          type: number
          description: HTTP status code
        type:
          type: string
          description: Error identifier
        message:
          type: string
          description: Human-readable error description
      required:
        - status
        - type
...